Rework ActiveSyntaxRewriter to operate on top of ConfiguredRegions
The operation that removes all inactive (and unparsed) regions from a
syntax tree is a two-pass algorithm. For clients that have already created
a ConfiguredRegions instance, let that serve as the first pass.
ConfiguredRegions now tracks the active clauses of any evaluated `#if`s,
and the ActiveSyntaxRewriter uses that information to guide its traversal
and rewrite rather than re-evaluating all of the clauses.

This required reworking the way in which we handle recursion in
ActiveSyntaxRewriter. Before, we were rewriting nodes "outside-in", but
doing so broke node identity and, therefore, lookup of the active #if
clauses. Instead, make sure we rewrite from the leaves up to the root.

Clients still have the option of removing inactive/unparsed code from a
syntax node given just a build configuration, in which case we'll first
create a ConfiguredRegions.